cache

+++ lib/Our/Cache.rakumod
@@ -1,8 +1,6 @@
 unit module Our::Cache:api<1>:auth<Mark Devine (mark@markdevine.com)>;
 
 use  Base64::Native;
-use  Compress::Bzip2;
-use  JSON::Fast;
 
 sub cache (Str:D :$meta!, Mu :$data, Str :$dir-prefix = $*PROGRAM.IO.basename, Instant :$expire-older-than) is export {
 
@@ -22,13 +20,17 @@ sub cache (Str:D :$meta!, Mu :$data, Str :$dir-prefix = $*PROGRAM.IO.basename, I
     my Str  $path       = $cache-dir ~ '/' ~ base64-encode($meta, :str);
 
     if $data {
-        spurt $path, compressToBlob(base64-encode(to-json($data)));
+        spurt $path, $data;
         $path.IO.chmod(0o600);
     }
     else {
         if "$path".IO.e {
-            unlink $path    if $expire-older-than && "$path".IO.modified < $expire-older-than;
-            return from-json(base64-decode(decompressToBlob(slurp($path, :bin))).decode);
+            if $expire-older-than && "$path".IO.modified < $expire-older-than {
+                unlink $path    if $expire-older-than && "$path".IO.modified < $expire-older-than;
+            }
+            else {
+                return slurp $path;
+            }
         }
     }
     return;
